+++ src/main/java/dev/yudin/entities/Course.java
@@ -1,18 +1,18 @@
 package dev.yudin.entities;
 
 public class Course {
-    private long id;
+    private int id;
     private String name;
     private String description;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
-    
-    public void setId(long id) {
+
+    public void setId(int id) {
         this.id = id;
     }
-    
+
     public String getName() {
         return name;
     }
